docker configfile added
