+++ conan/internal/runner/docker.py
@@ -1,8 +1,9 @@
+from collections import namedtuple
 import os
 import json
 import platform
 import shutil
-import itertools
+import yaml
 from conan.api.model import ListPattern
 from conan.api.output import Color, ConanOutput
 from conan.api.conan_api import ConfigAPI
@@ -11,6 +12,61 @@ from conan.internal.runner import RunnerException
 from conans.client.profile_loader import ProfileLoader
 from conans.errors import ConanException
 from conans.model.version import Version
+
+
+def config_parser(file_path):
+    '''
+    - image (str)
+    run:
+        - name -> name (str)
+        - containerEnv -> environment (dict)
+        - containerUser -> user (str or int)
+        - privileged -> privileged (boolean)
+        - capAdd -> cap_add (list)
+        - securityOpt -> security_opt (list)
+        - mounts -> volumes (dict)
+
+    dockerfile:
+        - build.dockerfile -> dockerfile (str) -> dockerfile.read()
+        - build.context -> path (str)
+        - build.args -> buildargs (dict)
+        build.options -> X (python extra params [extra_hosts, network_mode, ...])
+        - build.cacheFrom -> cache_from (list)
+    '''
+    Build = namedtuple('Build', ['dockerfile', 'build_context', 'build_args', 'cache_from'])
+    Run = namedtuple('Run', ['name', 'environment', 'user', 'privileged', 'cap_add', 'security_opt', 'volumes'])
+    Conf = namedtuple('Conf', ['image', 'build', 'run'])
+    if file_path:
+        def _instans_or_error(value, obj):
+            if value and (not isinstance(value, obj)):
+                raise Exception(f"{value} must be a {obj}")
+            return value
+        with open(file_path, 'r') as f:
+            runnerfile = yaml.safe_load(f)
+        return Conf(
+            image=_instans_or_error(runnerfile.get('image'), str),
+            build=Build(
+                dockerfile=_instans_or_error(runnerfile.get('build', {}).get('dockerfile'), str),
+                build_context=_instans_or_error(runnerfile.get('build', {}).get('build_context'), str),
+                build_args=_instans_or_error(runnerfile.get('build', {}).get('build_args'), dict),
+                cache_from=_instans_or_error(runnerfile.get('build', {}).get('cacheFrom'), list),
+            ),
+            run=Run(
+                name=_instans_or_error(runnerfile.get('run', {}).get('name'), str),
+                environment=_instans_or_error(runnerfile.get('run', {}).get('containerEnv'), dict),
+                user=_instans_or_error(runnerfile.get('run', {}).get('containerUser'), str),
+                privileged=_instans_or_error(runnerfile.get('run', {}).get('privileged'), bool),
+                cap_add=_instans_or_error(runnerfile.get('run', {}).get('capAdd'), list),
+                security_opt=_instans_or_error(runnerfile.get('run', {}).get('securityOpt'), dict),
+                volumes=_instans_or_error(runnerfile.get('run', {}).get('mounts'), dict),
+            )
+        )
+    else:
+        return Conf(
+            image=None,
+            build=Build(dockerfile=None, build_context=None, build_args=None, cache_from=None),
+            run=Run(name=None, environment=None, user=None, privileged=None, cap_add=None, security_opt=None, volumes=None)
+        )
 
 
 def docker_info(msg, error=False):
@@ -32,7 +88,7 @@ def list_patterns(cache_info):
 
 
 class DockerRunner:
-    def __init__(self, conan_api, command, profile, args, raw_args):
+    def __init__(self, conan_api, command, host_profile, build_profile, args, raw_args):
         import docker
         import docker.api.build
         try:
@@ -44,6 +100,7 @@ class DockerRunner:
                                  "\n - Check if docker is installed and running"
                                  "\n - Run 'pip install docker>=5.0.0, <=5.0.3'")
         self.conan_api = conan_api
+        self.build_profile = build_profile
         self.args = args
         self.abs_host_path = make_abs_path(args.path)
         if args.format:
@@ -68,15 +125,17 @@ class DockerRunner:
         self.command = ' '.join([f'conan {command}'] + [f'"{raw_arg}"' if ' ' in raw_arg else raw_arg for raw_arg in raw_args] + ['-f json > create.json'])
 
         # Container config
-        self.dockerfile = profile.runner.get('dockerfile')
-        self.docker_build_context = profile.runner.get('docker_build_context')
-        self.image = profile.runner.get('image')
+        # https://containers.dev/implementors/json_reference/
+        self.configfile = config_parser(host_profile.runner.get('configfile'))
+        self.dockerfile = host_profile.runner.get('dockerfile') or self.configfile.build.dockerfile
+        self.docker_build_context = host_profile.runner.get('build_context') or self.configfile.build.build_context
+        self.image = host_profile.runner.get('image') or self.configfile.image
         if not (self.dockerfile or self.image):
             raise ConanException("'dockerfile' or docker image name is needed")
         self.image = self.image or 'conan-runner-default'
-        self.name = f'conan-runner-{profile.runner.get("suffix", "docker")}'
-        self.remove = str(profile.runner.get('remove', 'false')).lower() == 'true'
-        self.cache = str(profile.runner.get('cache', 'clean'))
+        self.name = self.configfile.image or f'conan-runner-{host_profile.runner.get("suffix", "docker")}'
+        self.remove = str(host_profile.runner.get('remove', 'false')).lower() == 'true'
+        self.cache = str(host_profile.runner.get('cache', 'clean'))
         self.container = None
 
     def run(self):
@@ -94,6 +153,10 @@ class DockerRunner:
                 self.container = self.docker_client.containers.get(self.name)
                 self.container.start()
             else:
+                if self.configfile.run.environment:
+                    environment.update(self.configfile.run.environment)
+                if self.configfile.run.volumes:
+                    volumes.update(self.configfile.run.volumes)
                 docker_info('Creating the docker container')
                 self.container = self.docker_client.containers.run(
                     self.image,
@@ -101,6 +164,10 @@ class DockerRunner:
                     name=self.name,
                     volumes=volumes,
                     environment=environment,
+                    user=self.configfile.run.user,
+                    privileged=self.configfile.run.privileged,
+                    cap_add=self.configfile.run.cap_add,
+                    security_opt=self.configfile.run.security_opt,
                     detach=True,
                     auto_remove=False)
             docker_info(f'Container {self.name} running')
@@ -112,8 +179,10 @@ class DockerRunner:
             self.run_command(self.command)
             self.update_local_cache()
         except ConanException as e:
+            error = True
             raise e
         except RunnerException as e:
+            error = True
             raise ConanException(f'"{e.command}" inside docker fail'
                                  f'\n\nLast command output: {str(e.stdout_log)}')
         finally:
@@ -128,12 +197,18 @@ class DockerRunner:
     def build_image(self):
         dockerfile_file_path = self.dockerfile
         if os.path.isdir(self.dockerfile):
-            dockerfile_file_path = os.path.join(self.dockerfile, 'Dockerfile')    
+            dockerfile_file_path = os.path.join(self.dockerfile, 'Dockerfile')
         with open(dockerfile_file_path) as f:
             build_path = self.docker_build_context or os.path.dirname(dockerfile_file_path)
             ConanOutput().highlight(f"Dockerfile path: '{dockerfile_file_path}'")
             ConanOutput().highlight(f"Docker build context: '{build_path}'\n")
-            docker_build_logs = self.docker_api.build(path=build_path, dockerfile=f.read(), tag=self.image)
+            docker_build_logs = self.docker_api.build(
+                path=build_path,
+                dockerfile=f.read(),
+                tag=self.image,
+                buildargs=self.configfile.build.build_args,
+                cache_from=self.configfile.build.cache_from,
+            )
         for chunk in docker_build_logs:
                 for line in chunk.decode("utf-8").split('\r\n'):
                     if line:
